Fix delete character through My Characters
- move delete method into character list component
- edit character tile components
